+++ src/realm/alloc_slab.cpp
@@ -726,7 +726,7 @@ ref_type SlabAlloc::attach_file(const std::string& file_path, Config& cfg)
     // the session initiator. Another process may have the session initiator.
 
     m_file_mappings->m_file.open(path.c_str(), access, create, 0); // Throws
-    size_t size_without_considering_encryption = m_file_mappings->m_file.get_size();
+    size_t physical_file_size = m_file_mappings->m_file.get_size();
     if (cfg.encryption_key) {
         m_file_mappings->m_file.set_encryption_key(cfg.encryption_key);
     }
@@ -737,7 +737,7 @@ ref_type SlabAlloc::attach_file(const std::string& file_path, Config& cfg)
     // size_t.
     if (REALM_UNLIKELY(int_cast_with_overflow_detect(m_file_mappings->m_file.get_size(), size)))
         throw InvalidDatabase("Realm file too large", path);
-    if (cfg.encryption_key && size == 0 && size_without_considering_encryption != 0) {
+    if (cfg.encryption_key && size == 0 && physical_file_size != 0) {
         // The opened file holds data, but is so small it cannot have
         // been created with encryption
         throw std::runtime_error("Attempt to open unencrypted file with encryption key");