+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ### Bugfixes
 
-* Attemp to open a smal unencrypted realm file with an encryption key would
-  produce an empty encrypted realm file. Fixed by detecting the case and
+* Attempting to open a small unencrypted Realm file with an encryption key would
+  produce an empty encrypted Realm file. Fixed by detecting the case and
   throwing an exception.
   PR [#2645](https://github.com/realm/realm-core/pull/2645)